+++ src/components/Editor/Project/Project.js
@@ -30,7 +30,7 @@ import EditorInput from "../EditorInput/EditorInput";
 import NewFileModal from "../../Modals/NewFileModal";
 import ResizableWithHandle from "../../../utils/ResizableWithHandle";
 import { projContainer } from "../../../utils/containerQueries";
-import ProjectStatus from "../../ProjectStatus/ProjectStatus";
+import MobileProjectBar from "../../MobileProjectBar/MobileProjectBar";
 
 const Project = (props) => {
   const dispatch = useDispatch();
@@ -158,7 +158,7 @@ const Project = (props) => {
               <EditorInput />
             </ResizableWithHandle>
             <Output />
-            {isMobile ? <ProjectStatus /> : null}
+            {isMobile ? <MobileProjectBar /> : null}
           </div>
         </div>
       </div>